rewrite and add a lot of TS-related tests, make typing tests faster fix caching of TS response
