+++ package.json
@@ -152,7 +152,7 @@
         "start-web": "vscode-framework start --web",
         "build": "vscode-framework build && node buildTsPlugin.mjs prod",
         "watch-plugin": "node buildTsPlugin.mjs",
-        "integration-test": "tsc -p tsconfig.test.json && node testsOut/runTests.js",
+        "integration-test": "rimraf -rf ./testsOut/ && tsc -p tsconfig.test.json && node testsOut/runTests.js",
         "integration-test:watch": "chokidar \"test/integration/**\" -c \"pnpm integration-test\" --initial",
         "unit-test": "vitest --globals",
         "test": "vitest --globals --run && pnpm integration-test",
@@ -162,7 +162,8 @@
     "pnpm": {
         "overrides": {
             "esbuild": "^0.14.10",
-            "typescript-json-schema": "^0.53.0"
+            "typescript-json-schema": "^0.53.0",
+            "escape-string-regexp": "^4"
         },
         "neverBuiltDependencies": [
             "playwright"
@@ -174,6 +175,7 @@
         "@types/fs-extra": "^9.0.13",
         "@types/glob": "^7.2.0",
         "@types/jest": "^27.4.0",
+        "@types/lodash": "^4.14.187",
         "@types/mocha": "^9.1.0",
         "@types/node": "^16.11.25",
         "@types/path-browserify": "^1.0.0",
@@ -185,14 +187,16 @@
         "eslint-config-zardoy": "^0.2.14",
         "fs-extra": "^10.0.1",
         "glob": "^7.2.0",
+        "lodash": "^4.17.21",
         "mocha": "^9.2.0",
+        "rimraf": "^3.0.2",
         "tsm": "^2.2.1",
         "typescript": "^4.5.5",
         "vitest": "^0.8.1",
         "vscode-manifest": "^0.0.8"
     },
     "dependencies": {
-        "@zardoy/utils": "^0.0.9",
+        "@zardoy/utils": "^0.0.10",
         "@zardoy/vscode-utils": "^0.0.43",
         "chai": "^4.3.6",
         "chokidar-cli": "^3.0.0",
@@ -201,7 +205,6 @@
         "escape-string-regexp": "^5.0.0",
         "got": "^11.8.3",
         "jsonc-parser": "^3.0.0",
-        "lodash.orderby": "^4.6.0",
         "modify-json-file": "^1.2.2",
         "path-browserify": "^1.0.1",
         "prettier": "^2.5.1",
