+++ test/integration/suite/completion.test.ts
@@ -1,10 +1,11 @@
 import * as vscode from 'vscode'
 
 import { expect } from 'chai'
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment, @typescript-eslint/prefer-ts-expect-error
+import delay from 'delay'
+// eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error
 //@ts-ignore
 import type { Configuration } from '../../../src/configurationType'
-import { getFirstCompletionItem } from './utils'
+import { clearEditorText, getFirstCompletionItem } from './utils'
 
 const getFirstItemInsertText = async (pos: vscode.Position) => {
     const firstItem = await getFirstCompletionItem(pos)
@@ -52,6 +53,17 @@ describe('Basic copmletions', () => {
                             locations: ['fileStart'],
                         },
                     },
+                    {
+                        name: 'destruct3',
+                        body: 'const { $LAST } = $EXPR',
+                        when: {
+                            lineRegex: '(?<EXPR>(\\.?\\w)+)\\.(?<LAST>\\w+)\\.\\w*$',
+                            triggerCharacters: ['.', ''],
+                            languages: ['markdown'],
+                            locations: [],
+                        },
+                        replaceBeforeRegex: true,
+                    },
                 ]
                 await vscode.workspace.getConfiguration('betterSnippets').update(configKey, configValue, vscode.ConfigurationTarget.Global)
             })
@@ -68,21 +80,27 @@ describe('Basic copmletions', () => {
         expect: [new vscode.Position(1, 0), new vscode.Position(1, 3), new vscode.Position(1, contentLines[1]!.length)],
         dontExpect: [new vscode.Position(2, contentLines[2]!.length)],
     }
-    for (const [i, position] of positions.expect.entries())
+    for (const [i, position] of positions.expect.entries()) {
         it(`Expect snippet in position ${i}`, async () => {
             const snippet = await getFirstItemInsertText(position)
             expect(snippet).to.equal(expectedSnippet)
         })
+    }
 
-    for (const [i, position] of positions.dontExpect.entries())
+    for (const [i, position] of positions.dontExpect.entries()) {
         it(`Don't expect snippet in position ${i}`, async () => {
             const snippet = await getFirstItemInsertText(position)
             expect(snippet).to.not.equal(expectedSnippet)
         })
+    }
 
-    // it('Builtin em snippet should be on top', async () => {
-    //     await triggerFirstAtEnd()
-    //     await delay(500)
-    //     expect(activeEditor.getText()).to.equal(normalizeEol('codeblock'))
-    // })
+    it('replaceBeforeRegex + empty trigger character', async () => {
+        await clearEditorText(vscode.window.activeTextEditor!, 'test.a.b.de')
+        await vscode.commands.executeCommand('cursorEnd')
+        await delay(400)
+        await vscode.commands.executeCommand('editor.action.triggerSuggest')
+        await delay(800)
+        await vscode.commands.executeCommand('acceptSelectedSuggestion')
+        expect(document.getText()).to.equal('const { b } = test.a')
+    })
 })
